+++ pkg/config/snmpmetriccfg.go
@@ -1,6 +1,7 @@
 package config
 
 import (
+	"encoding/json"
 	"errors"
 	"fmt"
 	"regexp"
@@ -18,6 +19,42 @@ type MetricMultiMap struct {
 	Value interface{}
 }
 
+type ConversionMode int
+
+const (
+	FLOAT     ConversionMode = 0
+	INTEGER   ConversionMode = 1
+	UNSIGNED  ConversionMode = 2
+	STRING    ConversionMode = 3
+	BOOLEAN   ConversionMode = 4
+	HWADDRESS ConversionMode = 5
+	IPADDRESS ConversionMode = 6
+	NONE      ConversionMode = 5000
+)
+
+// GetString transform type to string
+func (c ConversionMode) GetString() string {
+	switch c {
+	case NONE:
+		return "NONE"
+	case FLOAT:
+		return "FLOAT"
+	case INTEGER:
+		return "INTEGER"
+	case UNSIGNED:
+		return "UNSIGNED INTEGER"
+	case STRING:
+		return "STRING"
+	case BOOLEAN:
+		return "BOOLEAN"
+	case HWADDRESS:
+		return "HARDWARE ADDRESS"
+	case IPADDRESS:
+		return "IP ADDRESS"
+	}
+	return ""
+}
+
 //SnmpMetricCfg Metric config
 type SnmpMetricCfg struct {
 	ID          string         `xorm:"'id' unique" binding:"Required"` //name of the key in the config array
@@ -28,9 +65,38 @@ type SnmpMetricCfg struct {
 	GetRate     bool           `xorm:"getrate"` //ony Valid with COUNTERS
 	Scale       float64        `xorm:"scale"`
 	Shift       float64        `xorm:"shift"`
-	IsTag       bool           `xorm:"'istag' default 0"` //Not Valid on  MULTISTRINGPARSER
-	ExtraData   string         `xorm:"extradata"`         //Only Valid with STRINGPARSER, MULTISTRINGPARSER, STRINGEVAL , BITS , BITSCHK, ENUM
-	Names       map[int]string `xorm:"-" json:"-"`        //BitString Name array
+	IsTag       bool           `xorm:"'istag' default 0"`      //Not Valid on  MULTISTRINGPARSER
+	ExtraData   string         `xorm:"extradata"`              //Only Valid with STRINGPARSER, MULTISTRINGPARSER, STRINGEVAL , BITS , BITSCHK, ENUM
+	Conversion  ConversionMode `xorm:"'conversion' default 0"` // Conversion will be always float for
+	Names       map[int]string `xorm:"-" json:"-"`             //BitString Name array
+}
+
+// MarshalJSON marshall
+func (m *SnmpMetricCfg) MarshalJSON() ([]byte, error) {
+	type Alias SnmpMetricCfg
+	return json.Marshal(&struct {
+		Conversion int `json:"conversion"`
+		*Alias
+	}{
+		Conversion: int(m.Conversion),
+		Alias:      (*Alias)(m),
+	})
+}
+
+// UnmarshalJSON unmarshall for conversion mode
+func (m *SnmpMetricCfg) UnmarshalJSON(data []byte) error {
+	type Alias SnmpMetricCfg
+	aux := &struct {
+		Conversion int `json:"Conversion"`
+		*Alias
+	}{
+		Alias: (*Alias)(m),
+	}
+	if err := json.Unmarshal(data, &aux); err != nil {
+		return err
+	}
+	m.Conversion = ConversionMode(aux.Conversion)
+	return nil
 }
 
 /*
@@ -142,6 +208,53 @@ func (m *SnmpMetricCfg) Init() error {
 		return fmt.Errorf("ExtraData not set in metric Config %s type  %s", m.ID, m.DataSrcType)
 	}
 	return nil
+}
+
+// GetValidConversions return Conversion Modes Array and the de default/sugested value beginning from 0
+func (m *SnmpMetricCfg) GetValidConversions() ([]ConversionMode, ConversionMode, error) {
+	switch m.DataSrcType {
+	case "INTEGER",
+		"Integer32",
+		"Gauge32",
+		"UInteger32",
+		"Unsigned32":
+		return []ConversionMode{FLOAT, INTEGER}, INTEGER, nil
+	case "Counter32",
+		"Counter64":
+		return []ConversionMode{FLOAT, INTEGER}, INTEGER, nil
+	case "COUNTER32",
+		"COUNTER64",
+		"COUNTERXX": //raw and cooked increment of Counter32
+		if m.GetRate == true {
+			return []ConversionMode{FLOAT, INTEGER}, FLOAT, nil
+		} else {
+			return []ConversionMode{FLOAT, INTEGER}, INTEGER, nil
+		}
+	case "TimeTicks", "TIMETICKS": //raw and cooked to second of timeticks
+		return []ConversionMode{FLOAT, INTEGER}, INTEGER, nil
+	case "BITSCHK":
+		return []ConversionMode{FLOAT, INTEGER, BOOLEAN}, BOOLEAN, nil
+	case "BITS": //no conversion  neeeded (not triggered)
+		return []ConversionMode{STRING}, STRING, nil
+	case "ENUM": //no conversion  neeeded (not triggered)
+		return []ConversionMode{STRING}, STRING, nil
+	case "OCTETSTRING": //no conversion  needed (not triggered)
+		return []ConversionMode{STRING, INTEGER}, STRING, nil
+	case "OID": //no conversion  neeeded (not triggered)
+		return []ConversionMode{STRING}, STRING, nil
+	case "HWADDR", "IpAddress": //no conversion  neeeded (not triggered)
+		return []ConversionMode{STRING}, STRING, nil
+	case "STRINGPARSER":
+		return []ConversionMode{FLOAT, INTEGER, BOOLEAN, STRING}, FLOAT, nil
+	case "MULTISTRINGPARSER": //no conversion  needed
+		return []ConversionMode{NONE}, NONE, nil
+	case "STRINGEVAL":
+		return []ConversionMode{FLOAT, INTEGER, BOOLEAN, STRING}, FLOAT, nil
+	case "CONDITIONEVAL": //not conversion will be triggered
+		return []ConversionMode{INTEGER}, INTEGER, nil
+	default:
+		return []ConversionMode{NONE}, NONE, errors.New("UnkNown DataSourceType:" + m.DataSrcType + " in metric Config " + m.ID)
+	}
 }
 
 // CheckEvalCfg : check evaluated expresion based in govaluate